app/SmartFoodChatbot/backend_ai/requirements.txt
# Web Framework nhanh nhất hiện nay
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
requests==2.31.0

# Google Gemini (AI miễn phí, mạnh mẽ)
google-generativeai==0.3.2

# Kết nối Supabase
supabase==2.3.0

# Thư viện xử lý dữ liệu cho tính năng Recommend (Nặng)
pandas==2.2.0
scikit-learn==1.4.0
numpy==1.26.3